requirements.txt
antlr4-python3-runtime==4.13.1
asttokens==2.4.1
comm==0.2.2
contourpy==1.2.1
coverage==7.4.4
cycler==0.12.1
debugpy==1.8.1
decorator==5.1.1
exceptiongroup==1.2.0
execnet==2.1.0
executing==2.0.1
fonttools==4.51.0
iniconfig==2.0.0
ipykernel==6.29.4
ipython==8.24.0
jedi==0.19.1
jupyter_client==8.6.1
jupyter_core==5.7.2
kiwisolver==1.4.5
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
packaging==24.0
parso==0.8.4
pluggy==1.4.0
prompt-toolkit==3.0.43
psutil==5.9.8
pure-eval==0.2.2
Pygments==2.17.2
pytest==8.1.1
pytest-cov==5.0.0
pytest-xdist==3.5.0
python-dateutil==2.9.0.post0
pyzmq==26.0.2
stack-data==0.6.3
tomli==2.0.1
tornado==6.4
traitlets==5.14.3
typing_extensions==4.11.0
wcwidth==0.2.13
# fichier rajouter 
# requirements.txt
streamlit
pandas
numpy
torch
datasets
faiss-cpu
transformers
sentence-transformers
# kagglehub       # si besoin
keybert         # Pour l’extraction de mots-clés
torchvision
